[FEAT] Filter button is now in both Favorites/Watchlist screen
